refactor(xpu): relocate FP8 weight implementation
Move MMWeightFp8IntelXpu into the Intel XPU platform ops and revert the unused Audio VAE weight-norm fusion.
